+++ src/nsidc/metgen/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "v1.11.0rc3"
+__version__ = "v1.12.0rc0"
 
 
 __all__ = [