+++ src/nsidc/metgen/cli.py
@@ -56,7 +56,7 @@ def info(config_filename):
     configuration = config.configuration(
         config.config_parser_factory(config_filename), {}
     )
-    metgen.init_logging()
+    metgen.init_logging(configuration)
     configuration.show()
 
 
@@ -82,7 +82,7 @@ def validate(config_filename, content_type):
     configuration = config.configuration(
         config.config_parser_factory(config_filename), {}
     )
-    metgen.init_logging()
+    metgen.init_logging(configuration)
     metgen.validate(configuration, content_type)
 
 
@@ -146,7 +146,7 @@ def process(config_filename, dry_run, env, number, write_cnm, overwrite):
         configuration = config.configuration(
             config.config_parser_factory(config_filename), overrides, env
         )
-        metgen.init_logging()
+        metgen.init_logging(configuration)
         configuration.show()
         config.validate(configuration)
         config.validate_spatial_source(configuration)